+++ src/ej_clusters/enums.py
@@ -1,17 +1,5 @@
-from django.db.models import TextChoices
+from django.db.models import IntegerChoices
 from django.utils.translation import gettext_lazy as _
-
-from boogie.fields import IntEnum
-from ej_conversations.enums import Choice
-
-
-class StereotypeVoteChoices(TextChoices):
-
-    agree = "agree"
-    disagree = "disagree"
-    skip = "skip"
-    __empty__ = "--------"
-
 
 FORM_CHOICE_MAP = {
     "1": "agree",
@@ -20,10 +8,8 @@ FORM_CHOICE_MAP = {
     "": "--------",
 }
 
-CHOICE_MAP = {"agree": Choice.AGREE, "disagree": Choice.DISAGREE, "skip": Choice.SKIP}
 
-
-class ClusterStatus(IntEnum):
+class ClusterStatus(IntegerChoices):
     PENDING_DATA = 0, _("Waiting for more data")
     ACTIVE = 1, _("Active")
     DISABLED = 2, _("Disabled")