+++ src/ej_clusters/models/stereotype.py
@@ -5,7 +5,6 @@ from django.conf import settings
 from django.db.models import OuterRef, Subquery
 from django.utils.translation import gettext_lazy as _
 
-from ej_clusters.enums import CHOICE_MAP
 from ej_conversations.enums import Choice
 from .querysets import StereotypeQuerySet
 from .stereotype_vote import StereotypeVote
@@ -86,7 +85,3 @@ class Stereotype(models.Model):
         return conversation.comments.filter(id__in=comment_ids).annotate(
             choice=Subquery(voted_subquery)
         )
-
-    @staticmethod
-    def get_choice_value(choice_name: str):
-        return CHOICE_MAP[choice_name]