+++ src/ej_conversations/enums.py
@@ -23,4 +23,27 @@ class Choice(models.IntegerChoices):
     SKIP = 0, _("Skip")
     AGREE = 1, _("Agree")
     DISAGREE = -1, _("Disagree")
-    __empty__ = _("(Unknown)")
+
+    @classmethod
+    def normalize(cls, choice):
+        """
+        Converts 'agree', 'skip' and 'disagree' to 1, 0 and -1.
+        It also converts '-1', '1' and '0' to -1, 1, and 0.
+        """
+
+        if choice is None or choice == "":
+            return None
+
+        if type(choice) == Choice:
+            return choice.value
+
+        if len(str(choice)) <= 2:
+            if int(choice) in cls:
+                return int(choice)
+            else:
+                raise ValueError
+
+        try:
+            return cls[choice.upper()].value
+        except KeyError as e:
+            raise e