+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.8.1] - Oct 01, 2024
+
+### Changed
+
+- Upgrades Django version to 4.2 .
+- Fixes Profile app migrations.
+- Refactoring ej_clusters app to use Django IntegerChoices instead of django-boogie EnumField.
+- [Web]: Fixes user's report perfomance issue.
+
+
 ## [3.8.0] - Sep 09, 2024
 
 ### Added